move AllToolsList into a file of its own
